Scraper/VPS_Scraper/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiohttp>=3.9.0
boto3>=1.34.0
shapely>=2.0.0
geopandas>=0.14.0
fiona>=1.9.0
geopy>=2.4.0
websockets>=12.0
